pyproject.toml
[project]
name = "ncar-caesar"
version = "0.1.0"
description = "CAESAR"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cartopy>=0.25.0",
    "netcdf4>=1.7.4",
    "polars>=1.38.1",
    "pyarrow>=23.0.1",
    "rdp>=0.8",
    "seaborn>=0.13.2",
    "xarray>=2026.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pillow>=11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "src/638-001", "src/638-038"]
